requirements.txt
fastapi==0.111.1
uvicorn==0.30.4
requests==2.32.3
fastapi-cors==0.0.6
pydantic==1.10.16
aiofiles==24.1.0
b2sdk==2.5.0
python-dotenv
mangum
fal_client
serpapi
google-search-results
httpx
googlesearch-python
beautifulsoup4
html5lib
openai
elevenlabs